+++ R/SimGeno.R
@@ -153,8 +153,8 @@ SimGeno <- function(Pedigree,
                     SnpError = 5e-4,
                     ErrorFV = function(E) c((E/2)^2, E-(E/2)^2, E/2),  # hom|hom, het|hom, hom|het
                     ErrorFM = NULL,
-					          ReturnStats = FALSE,
-					          quiet = FALSE)
+                    ReturnStats = FALSE,
+                    quiet = FALSE)
 {
   if (missing(Pedigree)) stop("please provide a pedigree to simulate from")
 
@@ -172,9 +172,9 @@ SimGeno <- function(Pedigree,
   params <- list('ParMis' = ParMis, 'MAF' = MAF,
                  'SnpError' = SnpError, 'CallRate' = CallRate)
   ValidLength <- list('ParMis' = c(1,2),
-  'MAF' = c(1, nSnp),
-  'SnpError' = c(1,3,nSnp,3*nSnp),
-  'CallRate' = c(1, nSnp))
+                      'MAF' = c(1, nSnp),
+                      'SnpError' = c(1,3,nSnp,3*nSnp),
+                      'CallRate' = c(1, nSnp))
 
   for (p in seq_along(params)) {
     if (is.null(params[[p]]) | all(is.na(params[[p]]))) {
@@ -210,7 +210,7 @@ SimGeno <- function(Pedigree,
   # check & prep ===
   Ped <- sequoia::PedPolish(Pedigree, ZeroToNA=TRUE)
   nInd <- nrow(Ped)
-  if (any(round(Q*nInd) %in% c(0,1)))  warning("some simulated SNPs have fixed alleles")
+  if (any(round(Q*nInd) %in% c(0,1)))  cli::cli_alert_warning("some simulated SNPs have fixed alleles")
 
 
   #================================
@@ -381,7 +381,7 @@ MkGenoErrors <- function(SGeno,
   #~~~~~~~~~
   if (any(CallRate <1)) {
     CRtype <- ifelse(length(CallRate)==1, "mean",
-                ifelse(!is.null(names(CallRate)), "Indiv", "SNP"))
+                     ifelse(!is.null(names(CallRate)), "Indiv", "SNP"))
     MisX <- matrix(FALSE, nInd, nSnp)
 
     if (CRtype == "Indiv") {
@@ -416,7 +416,7 @@ MkGenoErrors <- function(SGeno,
     return(list(GenoM = SGeno, Log = edit_log,
                 Counts_actual = table(factor(SGeno_orig, levels=c(-9,0,1,2)))))
   } else {
-  #~~~~~~~~~
+    #~~~~~~~~~
     return( SGeno )
   }
 }
@@ -437,20 +437,20 @@ MkGenoErrors <- function(SGeno,
 #' @keywords internal
 
 DoErrors <- function(SGeno, Act2Obs) {
-   dnames <- dimnames(SGeno)
-   # Generate random numbers to determine which SNPs are erroneous (r < p)
-   # random number generation by Fortran not allowed: F90 not always supported
-   # + may interfere with other pkgs
+  dnames <- dimnames(SGeno)
+  # Generate random numbers to determine which SNPs are erroneous (r < p)
+  # random number generation by Fortran not allowed: F90 not always supported
+  # + may interfere with other pkgs
 
-   randomV <- runif(n=nrow(SGeno)*ncol(SGeno), min=0, max=1)
+  randomV <- runif(n=nrow(SGeno)*ncol(SGeno), min=0, max=1)
 
-   TMP <- .Fortran(mkerrors,
-                   nind = as.integer(nrow(SGeno)),
-                   nsnp = as.integer(ncol(SGeno)),
-                   genofr = as.integer(SGeno),
-                   eprobfr = as.double(Act2Obs),
-                   randomv = as.double(randomV))
-   return( matrix(TMP$genofr, nrow(SGeno), ncol(SGeno), dimnames=dnames) )
+  TMP <- .Fortran(mkerrors,
+                  nind = as.integer(nrow(SGeno)),
+                  nsnp = as.integer(ncol(SGeno)),
+                  genofr = as.integer(SGeno),
+                  eprobfr = as.double(Act2Obs),
+                  randomv = as.double(randomV))
+  return( matrix(TMP$genofr, nrow(SGeno), ncol(SGeno), dimnames=dnames) )
 }
 
 
@@ -477,12 +477,14 @@ SelectNotSampled <- function(Ped, ParMis) {
 
   } else {
 
+    NotSampled <- NULL
     for (p in 1:2) {
       if (ParMis[p]>0) {
         IsParent <- which(Ped[,1] %in% Ped[,p+1])
       }
       if (round(length(IsParent)*ParMis[p]) > 0) {
-        NotSampled <- sample(IsParent, round(length(IsParent)*ParMis[p]), replace=FALSE)
+        NotSampled <- c(NotSampled,
+                        sample(IsParent, round(length(IsParent)*ParMis[p]), replace=FALSE))
       }
     }
   }